Refactor Input for use with or without children.
